Switch to lnkinfo for .lnk parsing, and support parsing relative paths
Relative paths are more likely to lead to a valid file, because it's very unlikely that a user will have specifically mapped their Wine drive partitions to match their Windows setup.
However, not all shortcuts have a relative path built in (none do if the target is on another partition), so most .lnk files will fail to generate an icon.